% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{analyze_strong_counts}
\alias{analyze_strong_counts}
\title{A helper method that allows for multi-threading of calculating of strong peptide counts
This is an internal method used by process_strong and requires
the conn list of connection information to be passed into it by way of
cluster import.  This method will loop through all genus_id values in in_set
and populate a temporary table that stores the percentage of organisms in the
genus that reference a given peptide, for all peptides referenced by the genus.}
\usage{
analyze_strong_counts(in_set, conn)
}
\arguments{
\item{in_set}{a vector of genus_id values from the database}

\item{conn}{database connection values to create a MySQL connection}
}
\value{
the in_set vector that was passed into the method
}
\description{
A helper method that allows for multi-threading of calculating of strong peptide counts
This is an internal method used by process_strong and requires
the conn list of connection information to be passed into it by way of
cluster import.  This method will loop through all genus_id values in in_set
and populate a temporary table that stores the percentage of organisms in the
genus that reference a given peptide, for all peptides referenced by the genus.
}
\author{
Dustin Crockett
}
