% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{build_database}
\alias{build_database}
\title{This method executes all stages of the building the database.  Including deleting the existing
database, building the data model, populating all kegg information and performing the post
upload steps.}
\usage{
build_database(
  conn,
  file_location,
  log_filename,
  large_storage_found = F,
  large_storage = "",
  max_peptide_length = 250,
  core_count = 2,
  split_size = 10,
  strong_tolerance = 0.053,
  max_count = 0
)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{file_location}{The location of the rData organism files}

\item{log_filename}{The location and file name of the log file}

\item{large_storage_found}{when True use the large_storage location for peptide_map when
false place peptide map with the rest of the database}

\item{large_storage}{The storage location of the large storage}

\item{max_peptide_length}{the maximum character length of peptides to be imported}

\item{core_count}{This is to total number cores to be used during parallel processing
it should be no greater than 1/2 of total cores}

\item{split_size}{This is size of list that should be passed to each execution of
analyze_strong_count.}

\item{strong_tolerance}{the percentage difference between organism coverages within a genus
of the two highest covered genus, for the highest covered genus to be
have a strong peptide}

\item{max_count}{Starting with 1 how many more organisms to process into the database}
}
\value{
T
}
\description{
\code{build_database} Goes through all steps of build the database with Kegg data.
}
\author{
Dustin Crockett
}
