% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\docType{data}
\name{conn_list}
\alias{conn_list}
\title{Deletes all information in the database in preperation of data repopulation}
\format{
An object of class \code{list} of length 5.
}
\usage{
conn_list
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{force}{When true don't prompt the user to verify if they want to delete the sample}
}
\value{
true when successful
}
\description{
\code{delete_database} Drop the schema in preperation of data repopulation.
}
\author{
Dustin Crockett

Daniel Vogel
}
\keyword{datasets}
