% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{create_datamodel}
\alias{create_datamodel}
\title{Creates all tables and database code needed for this package}
\usage{
create_datamodel(
  conn,
  large_storage_found = F,
  large_storage = NULL,
  max_peptide_length = 250
)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{large_storage_found}{when True use the large_storage location for peptide_map when
false place peptide map with the rest of the database}

\item{large_storage}{The storage location of the large storage}

\item{max_peptide_length}{the maximum character length of peptides to be imported}
}
\value{
true when successful
}
\description{
\code{create_datamodel} Creates all tables and database code needed for this package.
}
\author{
Dustin Crockett
}
