% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{delete_organism}
\alias{delete_organism}
\title{Removes all data associated with the organism.  If the organism is partially
inserted it may not properly delete the information.  Also if peptide
information is partially inserted it will not fully poplate in future loads
of the organism.}
\usage{
delete_organism(conn, organism_id)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{organism_id}{the database id for the organism}
}
\description{
\code{delete_organism} removes all data associated with the organism
}
\author{
Dustin Crockett
}
