% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Returns the jaccard index for all organisms}
\usage{
jaccard_index(conn)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}
}
\value{
a data.frame of with a kegg_id and jaccard indexes with all other organisms in matrix form
}
\description{
\code{jaccard_index} Returns the jaccard index for all organisms.
}
\author{
Dustin Crockett
}
