% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{optimize_database}
\alias{optimize_database}
\title{Executes multiple steps to optimize the database performance.  First the count of organism
for each peptide is calculated and saved to each peptide.  After that indexes are created
top improve query performance for sample processing.  Processing message are sent to
console out.}
\usage{
optimize_database(conn)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}
}
\value{
T
}
\description{
\code{optimize_database} Executes multiple optimizations to improve database performance.
}
\author{
Dustin Crockett
}
