% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{organism_info}
\alias{organism_info}
\title{Returns information about all organisms in the database}
\usage{
organism_info(conn)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}
}
\value{
a data.frame of , this is to be used to pivot other dataframes into a matrix
}
\description{
\code{organism_info} Returns the sample_Id and peptide_tag for each sample in a given sample set.
}
\author{
Dustin Crockett
}
