% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{overlap_coefficients}
\alias{overlap_coefficients}
\title{Returns the overlap coefficients for all organisms}
\usage{
overlap_coefficients(conn)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}
}
\value{
a data.frame of with a kegg_id and overlap coefficients with all other organisms in matrix form
}
\description{
\code{overlap_coefficients} Returns the overlap coefficients for all organisms.
}
\author{
Dustin Crockett
}
