% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{overlap_coefficients_strong}
\alias{overlap_coefficients_strong}
\title{Returns the strong peptide overlap coefficients for all organisms}
\usage{
overlap_coefficients_strong(conn)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}
}
\value{
a data.frame of with a kegg_id and overlap coefficients for strong peptides with all other organisms in matrix form
}
\description{
\code{overlap_coefficients_strong} Returns the strong peptide overlap coefficients for all organisms.
}
\author{
Dustin Crockett
}
