% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta.R
\name{parse_fasta}
\alias{parse_fasta}
\title{Parse a UniProt FASTA file}
\usage{
parse_fasta(
  input_file,
  output_dir,
  return_list = FALSE,
  hydrogen_mass = 1.00727646627
)
}
\arguments{
\item{input_file}{path to the input file}

\item{output_dir}{directory to write the output file, name will be the name
of the input file with the .RData extension}

\item{return_list}{Should the parsed FASTA file be returned as well as written
to a file? Defaults to FALSE.}

\item{hydrogen_mass}{mass of hydrogen to use when calculating neutral mass.
Default is 1.00727646627 which comes from
\href{https://physics.nist.gov/cgi-bin/cuu/Category?view=html&Atomic+and+nuclear.x=123&Atomic+and+nuclear.y=21}{NIST proton molar mass}}
}
\value{
Default is to return an RData object to the specified location.
}
\description{
Parse a UniProt FASTA file
}
\author{
Daniel T. Vogel
}
