% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{populate_first_step}
\alias{populate_first_step}
\title{A helper method that allows for multi-threading of organism populating
This is an internal method used by populate_organism and requires
the conn list of connection information to be passed into it by way of
cluster import.  This method can be called two different times with a
1 or a 2 as the in_set value to populate base tables needed for the next
step.}
\usage{
populate_first_step(in_set, conn)
}
\arguments{
\item{in_set}{a number that defines which step of logic to use}

\item{conn}{database connection values to create a MySQL connection}
}
\value{
string describing how long in seconds that it took for the part to complete including how many rows were inserted
}
\description{
A helper method that allows for multi-threading of organism populating
This is an internal method used by populate_organism and requires
the conn list of connection information to be passed into it by way of
cluster import.  This method can be called two different times with a
1 or a 2 as the in_set value to populate base tables needed for the next
step.
}
\author{
Dustin Crockett
}
