% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{populate_organism}
\alias{populate_organism}
\title{Populate the database with a given organism
This method is multi threaded.  First calling upload organism which uses two
threads to upload the organism data into the database.  Then this method
populate_first_step in two threads and finishes off by calling populate_second_step
in three threads.  The first step populates tables needed for the actions in
the seconds step.  The second step will not start until both threads in the first
step complete.  If this method gives errors verify that mysql server configuration
modifications have been made as documented.}
\usage{
populate_organism(conn, organism, max_peptide_length = 250)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{organism}{A list of data.frames that holds all information for the organism}

\item{max_peptide_length}{the maximum character length of peptides to be imported}
}
\value{
true when successful
}
\description{
Throughout the population process information messages are printed out.  At least
five lines of information is output (with word wrap most likely 8 lines).
}
\details{
\code{populate_organism} populates the database with the provided organism.
}
\author{
Dustin Crockett
}
