% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{populate_second_step}
\alias{populate_second_step}
\title{A helper method that allows for multi-threading of organism populating
This is an internal method used by populate_organism and requires
the has_pathways boolean variable, the has_modules boolean variable and,
the conn list of connection information to be passed into it by way of
cluster import.  This method can be called two different times with a
1 or a 2 as the in_set value to populate base tables needed for the next
step.}
\usage{
populate_second_step(in_set, conn, has_pathways = F, has_modules = F)
}
\arguments{
\item{in_set}{a number that defines which step of logic to use}

\item{conn}{database connection values to create a MySQL connection}

\item{has_pathways}{When T process the protein pathways that are already uploaded}

\item{has_modules}{When T process the protein modules that are already uploaded}
}
\value{
string describing how long in seconds that it took for the part to complete including how many rows were inserted
}
\description{
A helper method that allows for multi-threading of organism populating
This is an internal method used by populate_organism and requires
the has_pathways boolean variable, the has_modules boolean variable and,
the conn list of connection information to be passed into it by way of
cluster import.  This method can be called two different times with a
1 or a 2 as the in_set value to populate base tables needed for the next
step.
}
\author{
Dustin Crockett
}
