% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{populate_taxon}
\alias{populate_taxon}
\title{This method is executed after the Kegg database is fully loaded.  This method populates
taxon information into the database in adds taxon ids for each organism
A lineage table is also loaded that provides a full parent lineage for a taxon id}
\usage{
populate_taxon(conn, all_names, all_nodes, kegg_org_taxon, lineage_table)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{all_names}{a dataframe of all taxon names}

\item{all_nodes}{a dataframe of all taxons}

\item{kegg_org_taxon}{a dataframe with the taxon id for each organism in kegg}

\item{lineage_table}{a dataframe of the entire lineage for each taxon id}
}
\value{
T
}
\description{
\code{populate_taxon} Populates taxon information into a fully populated database.
}
\author{
Dustin Crockett
}
