% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{process_strong}
\alias{process_strong}
\title{Populate the database with strong peptides values
This method is multi threaded.  First creating a temporary table and then calling analyze_strong_counts
in parallel.  The number times analyze_strong_counts is called defined by the core_count variable,
the default value is 4.  Once analyze_strong_counts populates the
temporary table with the coverage of organisms by peptide and genus.  Next this temporary table
is agregated to create strong peptides records for all coverages that stand out from other coverages
by the tolerance provided in the tolerance variable, the
default value is .053.  Once that is done the temporary table is deleted.}
\usage{
process_strong(conn, core_count = 2, split_size = 10, tolerance = 0.053)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{core_count}{This is to total number cores to be used during parallel processing
it should be no greater than 1/2 of total cores}

\item{split_size}{This is size of list that should be passed to each execution of
analyze_strong_count.}

\item{tolerance}{the percentage difference between organism coverages within a genus
of the two highest covered genus, for the highest covered genus to be
have a strong peptide}
}
\description{
Throughout the population process information messages are printed out to show progress.
}
\details{
\code{process_strong} calculates which peptides that are strong for a genus, for all peptides.
}
\author{
Dustin Crockett
}
