% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{upload_organism}
\alias{upload_organism}
\title{Overwrites all data in the upload tables with the data from this list of dataframes
This method is multi threaded and will spawn off two threads that will upload
seperate parts of the organism into the database at the same time.}
\usage{
upload_organism(conn, organism, max_peptide_length)
}
\arguments{
\item{conn}{database connection values to create a MySQL connection}

\item{organism}{a list of dataframes that contains all informations for one organism in the KEGG}

\item{max_peptide_length}{the maximum character length of peptides to be imported}
}
\description{
\code{upload_organism} Creates all tables and database code needed for this package.
}
\author{
Dustin Crockett
}
