% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_database.linux.R
\name{upload_partial}
\alias{upload_partial}
\title{A helper method that allows for multi-threading of organism uploading
This is an internal method used by upload_organism and requires
the organism list of dataframes and conn list of connection information
to be passed into it through.  This method can be called two different
times with a 1 or a 2 as the in_set value to upload the two halves of
the organism.}
\usage{
upload_partial(in_set, conn, max_peptide_length, organism)
}
\arguments{
\item{in_set}{a number that defines which step of logic to use}

\item{conn}{database connection values to create a MySQL connection}

\item{max_peptide_length}{the maximum character length of peptides to be imported}

\item{organism}{a list of dataframes that contains all informations for one organism in the KEGG}
}
\value{
string describing how long in seconds that it took for the step to complete
}
\description{
A helper method that allows for multi-threading of organism uploading
This is an internal method used by upload_organism and requires
the organism list of dataframes and conn list of connection information
to be passed into it through.  This method can be called two different
times with a 1 or a 2 as the in_set value to upload the two halves of
the organism.
}
\author{
Dustin Crockett
}
